fix: combine UserClaimsDistribution components
